Add override_host for use in a container
When running minidlna in a docker container that is not exposed to
the host network, the advertised media must not use the private
non-routeable IP address because the consumer won't be able to find
it. This adds a config option to allow overriding the address we use
to advertise media, which if set will take precedence over the per-
interface address currently used.

This is useful for hosting multiple SSDP-advertised services on a
single box in docker containers with an SSDP proxy between that
private network and the physical one.